Gap reduction between tracks
The goal of the commit is to reduce the gap between tracks, especially
in a standard sequential playlist (not random). The current gap is
caused by:
1. the impossibility to determine accurately the end of the track since
   no `end` event is fired by Howler;
2. the RPC requests triggered in order to retrieve the next/previous
   track data;
3. the transcoding buffering delay;
4. the loading of the track in the web browser cache before playing.

Issue 1 is solved by using the `onend` function directly on the media
node. Although Howler doesn't fire the `end` event at the end of the
track, the `onend` method defined on the media node is called flawlessly
at the end of the track. This allows us to determine accurately the end
of the track.

Issues 2, 3 and 4 are solved by implementing a cache mechanism for media
nodes, i.e. Howler objects. Instead of unloading the Howler object
between tracks, we store it in a cache dictionary. Therefore, the data
is cached in the browser once at the first play of the track, then can
be reused in case it is played later on.

In order to ensure a smooth transition between tracks in a sequential
playlist, the media is cached 30 seconds before being played, i.e. 30
seconds before the end of the track being played. This delay should be
sufficient to ensure a proper transcoding buffering delay and track
loading in the web browser. This seems also reasonable when
normalization is activated, although the transcoding buffering delay is
much longer.